[Truffle] 2**63-1 ms is more than enough.
* It is about 292 471 208 years.

+++ truffle/src/main/java/org/jruby/truffle/nodes/core/MutexNodes.java
@@ -198,8 +198,7 @@ public abstract class MutexNodes {
 
         @Specialization
         public long sleep(RubyBasicObject mutex, NotProvided duration) {
-            // TODO: this should actually be "forever".
-            return doSleepMillis(mutex, Integer.MAX_VALUE);
+            return doSleepMillis(mutex, Long.MAX_VALUE);
         }
 
         @Specialization(guards = "isNil(duration)")
